refactor para
melhorar a criacao de comentarios
ajuste do ano (fix #263)
controle de acesso

+++ app/Http/Controllers/ArquivoController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Arquivo;
+use App\Models\Chamado;
 use App\Models\Comentario;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\URL;
 
 class ArquivoController extends Controller
 {
@@ -15,9 +18,7 @@ class ArquivoController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-        //
-    }
+    {}
 
     /**
      * Show the form for creating a new resource.
@@ -25,9 +26,7 @@ class ArquivoController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {
-        //
-    }
+    {}
 
     /**
      * Store a newly created resource in storage.
@@ -38,46 +37,48 @@ class ArquivoController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'arquivo.*' => 'required|mimes:jpeg,jpg,png,pdf|max:'.config('chamados.upload_max_filesize'),
+            'arquivo.*' => 'required|mimes:jpeg,jpg,png,pdf|max:' . config('chamados.upload_max_filesize'),
             'chamado_id' => 'required|integer|exists:chamados,id',
         ]);
+
+        $chamado = Chamado::find($request->chamado_id);
+        $this->authorize('chamados.update', $chamado);
+
         $fotos_nome = [];
-        $chamado_id = 0;
+        $arquivos_nome = [];
         foreach ($request->arquivo as $arq) {
             $arquivo = new Arquivo;
-            $arquivo->chamado_id = $request->chamado_id;
-            $chamado_id = $arquivo->chamado_id;
+            $arquivo->chamado_id = $chamado->id;
             $arquivo->user_id = \Auth::user()->id;
             $arquivo->nome_original = $arq->getClientOriginalName();
-            $arquivo->caminho = $arq->store('./arquivos/' . date("Y"));
+            $arquivo->caminho = $arq->store('./arquivos/' . $chamado->created_at->year);
             $arquivo->mimeType = $arq->getClientMimeType();
             $arquivo->save();
             if (preg_match('/jpeg|jpg|png/', $arquivo->mimeType)) {
                 array_push($fotos_nome, $arquivo->nome_original);
             } else {
-                Comentario::criar([
-                    'user_id' => \Auth::user()->id,
-                    'chamado_id' => $arquivo->chamado_id,
-                    'comentario' => 'O arquivo ' . $arquivo->nome_original . ' foi adicionado.',
-                    'tipo' => 'system',
-                ]);
+                array_push($arquivos_nome, $arquivo->nome_original);
             }
         }
-        if (count($fotos_nome) > 0) {
-            $comentario = (count($fotos_nome) > 1)
-            ? 'As imagens ' . implode(", ", $fotos_nome) . ' foram adicionadas'
-            : 'A imagem ' . $fotos_nome[0] . ' foi adicionada';
 
-            Comentario::criar([
-                'user_id' => \Auth::user()->id,
-                'chamado_id' => $arquivo->chamado_id,
-                'comentario' => $comentario,
-                'tipo' => 'system',
-            ]);
+        # arquivos
+        if (count($arquivos_nome) == 1) {
+            Comentario::criarSystem($chamado, 'O arquivo ' . $arquivos_nome[0] . ' foi adicionado.');
         }
-        $request->session()->flash('alert-success', 'Arquivo(s) adicionado(s) com sucesso!');
+        if (count($arquivos_nome) > 1) {
+            Comentario::criarSystem($chamado, 'Os arquivos ' . implode(', ', $arquivos_nome) . ' foram adicionados.');
+        }
 
-        return back();
+        # imagens
+        if (count($fotos_nome) == 1) {
+            Comentario::criarSystem($chamado, 'A imagem ' . $fotos_nome[0] . ' foi adicionada');
+        }
+        if (count($fotos_nome) > 1) {
+            Comentario::criarSystem($chamado, 'As imagens ' . implode(", ", $fotos_nome) . ' foram adicionadas');
+        }
+
+        $request->session()->flash('alert-success', 'Arquivo(s) adicionado(s) com sucesso!');
+        return Redirect::to(URL::previous() . "#card_arquivos");
     }
 
     /**
@@ -88,6 +89,8 @@ class ArquivoController extends Controller
      */
     public function show(Arquivo $arquivo)
     {
+        $this->authorize('chamados.view', $arquivo->chamado);
+
         return Storage::download($arquivo->caminho, $arquivo->nome_original);
     }
 
@@ -98,9 +101,7 @@ class ArquivoController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Arquivo $arquivo)
-    {
-        //
-    }
+    {}
 
     /**
      * Update the specified resource in storage.
@@ -111,26 +112,23 @@ class ArquivoController extends Controller
      */
     public function update(Request $request, Arquivo $arquivo)
     {
-        $request->validate([
-            'nome_arquivo' => 'required',
-        ],
-            [
-                'nome_arquivo.required' => 'O nome do arquivo é obrigatório!',
-            ]);
+        $this->authorize('chamados.update', $arquivo->chamado);
+
+        $request->validate(
+            ['nome_arquivo' => 'required'],
+            ['nome_arquivo.required' => 'O nome do arquivo é obrigatório!']
+        );
         $nome_antigo = $arquivo->nome_original;
-        $arquivo->nome_original = $request->nome_arquivo;
-        if (substr($arquivo->nome_original, -4) != '.pdf') {
-            $arquivo->nome_original .= '.pdf';
-        }
+        $arquivo->nome_original = $request->nome_arquivo . '.pdf';
         $arquivo->update();
-        Comentario::criar([
-            'user_id' => \Auth::user()->id,
-            'chamado_id' => $arquivo->chamado_id,
-            'comentario' => 'O arquivo ' . $nome_antigo . ' foi renomeado para ' . $request->nome_arquivo . '.',
-            'tipo' => 'system',
-        ]);
+
+        Comentario::criarSystem(
+            $arquivo->chamado,
+            'O arquivo ' . $nome_antigo . ' foi renomeado para ' . $arquivo->nome_original . '.'
+        );
+
         request()->session()->flash('alert-success', 'Arquivo renomeado com sucesso!');
-        return back();
+        return Redirect::to(URL::previous() . "#card_arquivos");
     }
 
     /**
@@ -141,18 +139,15 @@ class ArquivoController extends Controller
      */
     public function destroy(Request $request, Arquivo $arquivo)
     {
+        $this->authorize('chamados.update', $arquivo->chamado);
+
         $arquivo->delete();
         $comentario = preg_match('/jpeg|jpg|png/', $arquivo->mimeType)
         ? 'A imagem ' . $arquivo->nome_original . ' foi excluída'
         : 'O arquivo ' . $arquivo->nome_original . ' foi excluído';
-        Comentario::criar([
-            'user_id' => \Auth::user()->id,
-            'chamado_id' => $arquivo->chamado_id,
-            'comentario' => $comentario,
-            'tipo' => 'system',
-        ]);
-        $request->session()->flash('alert-success', $comentario . ' com sucesso!');
+        Comentario::criarSystem($arquivo->chamado, $comentario);
 
-        return back();
+        $request->session()->flash('alert-success', $comentario . ' com sucesso!');
+        return Redirect::to(URL::previous() . "#card_arquivos");
     }
 }
